carregar_energia_api/carregar_energia_api/requirements.txt
asgiref == 3.8.1
CacheControl == 0.14.1
cachetools== 5.5.0
certifi == 2024.8.30
cffi == 1.17.1
charset-normalizer == 3.4.0
cryptography == 44.0.0
Django == 5.1.4
django-cors-headers == 4.6.0
django-filter == 24.3
djangorestframework == 3.15.2
firebase-admin == 6.6.0
google-api-core == 2.23.0
google-api-python-client == 2.154.0
google-auth == 2.36.0
google-auth-httplib2 == 0.2.0
google-cloud-core == 2.4.1
google-cloud-firestore == 2.19.0
google-cloud-storage == 2.19.0
google-crc32c == 1.6.0
google-resumable-media == 2.7.2
googleapis-common-protos == 1.66.0
grpcio == 1.68.1
grpcio-status == 1.68.1
httplib2 == 0.22.0
idna == 3.10
Markdown == 3.7
msgpack == 1.1.0
pip == 22.0.2
proto-plus == 1.25.0
protobuf == 5.29.1
pyasn1 == 0.6.1
pyasn1_modules == 0.4.1
pycparser == 2.22
PyJWT == 2.10.1
pyparsing == 3.2.0
requests == 2.32.3
rsa == 4.9
setuptools == 59.6.0
sqlparse == 0.5.2
typing_extensions == 4.12.2
uritemplate == 4.1.1
urllib3 == 2.2.3